requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
openai>=1.6.0
matplotlib>=3.7.0
seaborn>=0.12.0
hdbscan>=0.8.30
umap-learn>=0.5.0
tqdm>=4.66.0
python-dotenv>=1.0.0 